+++ hashtagmap-admin/src/main/java/com/songpapeople/hashtagmap/kakao/service/KakaoScheduleCommandService.java
@@ -2,8 +2,11 @@ package com.songpapeople.hashtagmap.kakao.service;
 
 import com.songpapeople.hashtagmap.exception.AdminErrorCode;
 import com.songpapeople.hashtagmap.exception.AdminException;
+import com.songpapeople.hashtagmap.kakao.schedule.PeriodHistory;
+import com.songpapeople.hashtagmap.kakao.schedule.PeriodHistoryRepository;
 import com.songpapeople.hashtagmap.kakao.schedule.model.Schedule;
 import com.songpapeople.hashtagmap.kakao.schedule.repository.ScheduleRepository;
+import com.songpapeople.hashtagmap.response.CustomResponse;
 import com.songpapeople.hashtagmap.scheduler.domain.KakaoScheduler;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,6 +15,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 public class KakaoScheduleCommandService {
+    private final PeriodHistoryRepository historyRepository;
     private final ScheduleRepository scheduleRepository;
     private final KakaoScheduler kakaoScheduler;
 
@@ -33,4 +37,13 @@ public class KakaoScheduleCommandService {
                         String.format("스케쥴러(%s)가 존재하지 않습니다.", name)
                 ));
     }
+
+    @Transactional
+    public CustomResponse<Void> changeSchedulePeriod(String expression) {
+        kakaoScheduler.changePeriod(expression);
+        historyRepository.save(new PeriodHistory(expression));
+
+        return CustomResponse.empty();
+    }
+
 }
