+++ hashtagmap-admin/src/test/java/com/songpapeople/hashtagmap/service/KakaoServiceTest.java
@@ -1,6 +1,8 @@
 package com.songpapeople.hashtagmap.service;
 
 import com.songpapeople.hashtagmap.kakao.schedule.PeriodHistoryRepository;
+import com.songpapeople.hashtagmap.kakao.schedule.repository.ScheduleRepository;
+import com.songpapeople.hashtagmap.kakao.service.KakaoScheduleCommandService;
 import com.songpapeople.hashtagmap.scheduler.domain.CronPeriod;
 import com.songpapeople.hashtagmap.scheduler.domain.KakaoScheduler;
 import org.junit.jupiter.api.BeforeEach;
@@ -16,15 +18,18 @@ import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 class KakaoServiceTest {
-    private KakaoService kakaoService;
+    private KakaoScheduleCommandService kakaoService;
 
     @Mock
     private PeriodHistoryRepository periodHistoryRepository;
 
+    @Mock
+    private ScheduleRepository scheduleRepository;
+
     @BeforeEach
     public void setUp() {
         KakaoScheduler kakaoScheduler = new KakaoScheduler(new Thread(), new CronPeriod("0 0 * * * ?"));
-        kakaoService = new KakaoService(kakaoScheduler, periodHistoryRepository);
+        kakaoService = new KakaoScheduleCommandService(periodHistoryRepository, scheduleRepository, kakaoScheduler);
     }
 
     @DisplayName("카카오 스케줄러 주기 변경")
